feat: restructured folder
